pyproject.toml
[project]
name = "lightonocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.0",
    "pypdfium2>=5.3.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.0.0",
]

[project.scripts]
lighton-ocr = "lightonocr.run_ocr:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
